==> 5 - full_embeddings_optimized.ipynb <==
import numpy as np

# Replace 'embeddings_with_ids_batch_0.npz' with the actual filename you want to read
filename = "embeddings/embeddings_with_ids_batch_0.npz"
data = np.load(filename)

import sys
# Extract the IDs and embeddings
loaded_ids = data['ids']
loaded_embeddings = data['embeddings']

n_bytes = sys.getsizeof(loaded_embeddings)

n_bytes  / (1024 * 1024) *3000
