==> 6 - apply_ml_model.ipynb <==
import pickle

# Load the XGBoost model
with open('exports/model/xgboost_model.pkl', 'rb') as f:
    model = pickle.load(f)

# Load the label encoder
with open('exports/model/label_encoder.pkl', 'rb') as f:
    label_encoder = pickle.load(f)

import pandas as pd

df_topics = pd.read_csv('exports/topics.csv', index_col=[0])
df_topics

import numpy as np
import glob
from tqdm import tqdm

files = glob.glob('embeddings/*') 
#files = files[:100]

df_final_list=[]
for filename in tqdm(files):
    # Replace 'embeddings_with_ids_batch_0.npz' with the actual filename you want to read
    #filename = "embeddings/embeddings_with_ids_batch_0.npz"
    data = np.load(filename)
    loaded_embeddings = data['embeddings']
    ids =  data['ids']


    predicted_y = model.predict_proba(loaded_embeddings)
    class_labels_in_order = label_encoder.inverse_transform(model.classes_)
    df_predicted = pd.DataFrame(predicted_y, columns = class_labels_in_order)
    df_predicted['doc_id'] = ids
    df_predicted = df_predicted.set_index('doc_id')
    df_final_list.append(df_predicted)

df_final = pd.concat([x for x in df_final_list])

df_final.to_csv('exports/final_predictions_lemonde.csv')
